requirements.txt
minimalmodbus
paho-mqtt